+++ R/controlFunctions.R
@@ -9,9 +9,11 @@
 #' @param midp logical, use mid-p values? Currently only permitted
 #' where `tsmethod = "central"`.
 #' @param tsmethod `character` giving two-sided method
+#' @param ... Currently ignored..
 #' @export
 dobsonControl <- function(midp = FALSE,
-                          tsmethod = c("central", "minlike", "blaker")){
+                          tsmethod = c("central", "minlike", "blaker"),
+                          ...){
   tsmethod <- match.arg(tsmethod)
   list(midp = midp, tsmethod = tsmethod)
 }
@@ -22,12 +24,12 @@ dobsonControl <- function(midp = FALSE,
 #' of the MLE.
 #'
 #' @param trans Transformation apply to the MLE distribution. 
-#'
+#' @param ... Currently ignored.
 #' @return
 #' A list with values
 #' \item{\code{trans}}{}
 #' @export
-asymptoticControl <- function(trans = c("none", "log")){
+asymptoticControl <- function(trans = c("none", "log"), ...){
   trans <- match.arg(trans)
   list(trans = trans)
 }
@@ -42,6 +44,7 @@ asymptoticControl <- function(trans = c("none", "log")){
 #' @param wmtype type of modification for the Gamma confidence interval.
 #' @param unirootTolFactor tolerance factor for uniroot where `midp = TRUE`
 #' and `nmc = 0`.
+#' @param ... Currently ignored.
 #' @return
 #' A list of arguments to pass to \code{\link[asht]{wspoissonTest}}.
 #' If `midp = TRUE`, with values
@@ -54,7 +57,7 @@ asymptoticControl <- function(trans = c("none", "log")){
 #' @export
 gammaControl <- function(midp = FALSE, nmc = 0,
                          wmtype = c("max", "mean", "minmaxavg", "tcz"),
-                         unirootTolFactor = 1e-06){
+                         unirootTolFactor = 1e-06, ...){
   wmtype <- match.arg(wmtype)
   if (midp) {
     RVAL <- list(midp = midp, nmc = nmc,
@@ -67,12 +70,31 @@ gammaControl <- function(midp = FALSE, nmc = 0,
 
 #' Control Function for Beta Method for Confidence Intervals
 #'
-#' @description In the future, this may provide arguments
-#' that will permit modifications to the Beta method.
+#' @description Modification to the Beta method. The options are `"none"`
+#' or the same modifications as applied to the Gamma Method 
+#' (see \code{\link{gammaControl}}) are implemented. `wmtype="none"` and
+#' `wmtype="tcz"` have been investigated by Tiwari et al (2006) and
+#' Ng et al (2008).
+#' @references 
+#' Tiwari RC, Clegg LX, & Zou Z (2006). 'Efficient interval estimation 
+#' for age-adjusted cancer rates.' 
+#' *Statistical Methods in Medical Research* **15**: 547--569. 
+#' \doi{doi:10.1177/0962280206070621}
+#' 
+#' Ng HKT, Filardo, G & Zheng G (2008). 'Confidence interval estimating 
+#' procedures for standardized incidence rates.' 
+#' *Computational Statistics and Data Analysis* **52** 3501--3516.
+#' \doi{doi:10.1016/j.csda.2007.11.004}
+#' @param wmtype `character` type of modification to the Beta Confidence
+#' Interval
+#' @param ... Currently ignored.
 #'
-#' @param ... Currently ignored
-#'
-#' @return an empty list
+#' @return a list with values
+#' \item{wmtype}{modificator to Beta Confidence Interval to implement}
 #'
 #' @export
-betaControl <- function(...) list()
+betaControl <- function(wmtype = c("none", "tcz", "max", "mean",
+                                   "minmaxavg"), ...) {
+  wmtype <- match.arg(wmtype)
+  list(wmtype = wmtype)
+}