+++ NEWS.md
@@ -7,5 +7,4 @@
 - Bug Fixes on alternative = "less" or "greater"
 - Improved documentation formatting
 - Example data and vignette added
-
-
+- wmtype modifications added for Beta methods